feat: add github username as command line arg
To get around the issue of the having the person who creates the pull request included in the reviewers, the username of the person creating the pull request is now a required command line argument. This will filter their username out of the list of reviewers before making the request to github for creating the pull request.